+++ CHANGELOG.md
@@ -8,5 +8,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* number secret lead to encrypt failed ([f36df5b](https://github.com/alphawq/vuepress-plugin-posts-encrypt/commit/f36df5bd29524dd04d9ba6ce99fb7e3fafa30534))
-* **vuepress-plugin-posts-encrypt:** number secret lead to encrypt failed ([b893eb9](https://github.com/alphawq/vuepress-plugin-posts-encrypt/commit/b893eb9bbbb543b465edd192954e5ec29610ddcd))
+* number secret lead to encrypt failed ([018bb22](https://github.com/alphawq/mono-blog/commit/018bb22d9fe650b4f916159b4881969390e2ae44))
+* number secret lead to encrypt failed ([f36df5b](https://github.com/alphawq/mono-blog/commit/f36df5bd29524dd04d9ba6ce99fb7e3fafa30534))
+* **vuepress-plugin-posts-encrypt:** number secret lead to encrypt failed ([b893eb9](https://github.com/alphawq/mono-blog/commit/b893eb9bbbb543b465edd192954e5ec29610ddcd))